+++ server/docker/node/Dockerfile
@@ -8,12 +8,12 @@ WORKDIR /app
 ENV PATH /app/node_modules/.bin:$PATH
 
 # install app dependencies
-COPY package.json ./
+COPY ./client/package.json ./
 RUN npm install --silent
 RUN npm install react-scripts@3.4.1 -g --silent
 
 # add app
-COPY . ./
+#COPY . ./
 
 # start app
-CMD ["npm", "start"]
+#CMD ["npm", "start"]
